ai-personal-assistant/requirements.txt

openai
python-dotenv
langchain
langchain-openai

